feat: aPI機能実装＆GoogleAnalyticsアカウント情報登録機能実装

+++ articles/views.py
@@ -12,7 +12,7 @@ from accounts.models import User
 from articles.forms import UpdateArticleForm, AddArticleForm
 from projects.forms import KeywordForm
 
-from articles.management.commands.utils.get_specific_analytics_data import get_analytics_data_eq
+from articles.management.commands.utils.get_specific_analytics_data import validate_account_info, get_analytics_data_eq
 from projects.management.commands.utils.generate_keywords_api import generate_keywords_api
 
 from urllib.parse import urlparse
@@ -24,38 +24,6 @@ import re
 @login_required
 def top(request, project_id):
     current = Project.objects.get(id=project_id)
-    is_error = False
-    if request.method == 'POST':
-        form = AddArticleForm(request.POST)
-        path = request.POST.get('path')
-        if re.search(r'//', path):
-            o = urlparse(path)
-            host = o.hostname
-            path = o.path
-            if not host == current.website.domain:
-                is_error = True
-                messages.error(request, 'プロジェクトのドメインと一致しません')
-            elif Article.objects.get_or_none(project=current,path=path):
-                is_error = True
-                messages.error(request, '既に登録済みの記事です')
-        if form.is_valid() and not is_error:
-            article = form.save(commit=False)
-            article.path = path
-            article.project = current
-            article.person_in_charge = request.user
-            article.created_by = request.user
-            article.save()
-            res = get_analytics_data_eq(path, 6)
-            for data in reversed(res):
-                ArticleAnalytics.objects.create(
-                    article=article,
-                    date=data['date'],
-                    session=data['session'],
-                    conversion=data['cv'],
-                    conversion_rate=data['cvr'],
-                )
-            messages.success(request, '登録に成功しました')
-        return redirect(top, project_id=project_id)
     contents = []
     articles = Article.objects.filter(project=current).all()
     for article in articles:
@@ -80,7 +48,7 @@ def top(request, project_id):
                 kdata = {
                     "keyword": k.keyword,
                     "volume": k.volume,
-                    "ranking": ['-','-','-','-','-'],
+                    "ranking": [0,0,0,0,0],
                     "date": ['-','-','-','-','-'],
                     "right_wrong": "-",
                     "ranking_page": "-"
@@ -106,12 +74,67 @@ def top(request, project_id):
 
 @owner_check
 @login_required
+def add(request, project_id):
+    current = Project.objects.get(id=project_id)
+    is_error = False
+    if request.method == 'POST':
+        form = AddArticleForm(request.POST)
+        path = request.POST.get('path')
+        if re.search(r'//', path):
+            o = urlparse(path)
+            host = o.hostname
+            path = o.path
+            if not host == current.website.domain:
+                is_error = True
+                messages.error(request, 'プロジェクトのドメインと一致しません')
+            elif Article.objects.get_or_none(project=current,path=path):
+                is_error = True
+                messages.error(request, '既に登録済みの記事です')
+        if form.is_valid() and not is_error:
+            article = form.save(commit=False)
+            article.path = path
+            article.project = current
+            article.person_in_charge = request.user
+            article.created_by = request.user
+            article.save()
+            account_info = [current.account_id, current.property_id, current.view_id]
+            status_code = validate_account_info(account_info)
+            if status_code == 0:
+                res = get_analytics_data_eq(path, 6, account_info)
+                for data in reversed(res):
+                    ArticleAnalytics.objects.create(
+                        article=article,
+                        date=data['date'],
+                        session=data['session'],
+                        conversion=data['cv'],
+                        conversion_rate=data['cvr'],
+                    )
+            messages.success(request, '登録に成功しました')
+    return redirect(top, project_id=project_id)
+
+
+@owner_check
+@login_required
 def settings(request, project_id, article_id):
     current = Article.objects.get(id=article_id)
+    projects = request.user.members_projects.all()
+    members = current.project.members.all()
+    context = {
+        'current': current,
+        'projects': projects,
+        'members': members,
+        'keywords': current.keywords.all()
+    }
+    return render(request, 'articles/detail.html', context)
+
+
+@owner_check
+@login_required
+def update(request, project_id, article_id):
+    current = Article.objects.get(id=article_id)
     if request.method == 'POST':
-        aform = UpdateArticleForm(request.POST)
-        kform = KeywordForm(request.POST)
-        if aform.is_valid():
+        form = UpdateArticleForm(request.POST)
+        if form.is_valid():
             current.project = Project.objects.get(id=request.POST.get('project'))
             path = request.POST.get('path')
             if re.search(r'//', path):
@@ -124,8 +147,16 @@ def settings(request, project_id, article_id):
             current.category = request.POST.get('category')
             current.save()
             messages.success(request, "設定を更新しました")
-            return redirect(settings, project_id=project_id, article_id=article_id)
-        elif kform.is_valid(): 
+    return redirect(settings, project_id=project_id, article_id=article_id)
+
+
+@owner_check
+@login_required
+def add_keywords(request, project_id, article_id):
+    current = Article.objects.get(id=article_id)
+    if request.method == 'POST':
+        form = KeywordForm(request.POST)
+        if form.is_valid(): 
             keywords = request.POST.get('keyword')
             lst = keywords.split('\n')
             keyword = []
@@ -181,30 +212,26 @@ def settings(request, project_id, article_id):
                         ) 
                         wk.competitors.add(current.project)
             messages.success(request, 'キーワードの登録に成功しました')
-        elif request.POST.get('delete'):
-            current.delete()
-            messages.success(request, '記事を削除しました')
-            return redirect(top, project_id=project_id)
-        else:
-            messages.error(request, 'キーワードの登録に失敗しました')
-        return redirect(settings, project_id=project_id, article_id=article_id)
-    projects = request.user.members_projects.all()
-    members = current.project.members.all()
-    context = {
-        'current': current,
-        'projects': projects,
-        'members': members,
-        'keywords': current.keywords.all()
-    }
-    return render(request, 'articles/detail.html', context)
+    return redirect(settings, project_id=project_id, article_id=article_id)
 
 
 @owner_check
 @login_required
-def remove_keywords(request, project_id, article_id):
+def delete_keywords(request, project_id, article_id):
     current = Article.objects.get(id=article_id)
-    if request.POST.get('keyword'):
+    if request.method == 'POST' and request.POST.get('keyword'):
         for k in request.POST.getlist('keyword'):
             keyword = Keyword.objects.get(id=k)
             current.keywords.remove(keyword)
+    return redirect(settings, project_id=project_id, article_id=article_id)
+
+
+@owner_check
+@login_required
+def delete(request, project_id, article_id):
+    current = Article.objects.get(id=article_id)
+    if request.method == 'POST' and request.POST.get('delete'):
+        current.delete()
+        messages.success(request, '記事を削除しました')
+        return redirect(top, project_id=project_id)
     return redirect(settings, project_id=project_id, article_id=article_id)
